LiteLLM/mcp-server-test/requirements.txt
PyGithub
fastapi
uvicorn[standard]
python-dotenv
mcp

